Rename 'components/ConfirmedCasesDetailsTable.vue' to 'components/index/CardsMonitoring/ConfirmedCasesDetails/StackedTable.vue'
