+++ components/index/CardsMonitoring/ConfirmedCasesDetails/StackedTable.vue
